M6 Demo

+++ app/src/main/java/ajacks/cs2340/edu/gatech/cs2340_ajacks/model/Coordinates.java
@@ -1,23 +1,23 @@
 package ajacks.cs2340.edu.gatech.cs2340_ajacks.model;
 
 /**
+ * Stores coordinate information in a nice wrapper.
  * Created by Alonzo on 10/4/2017.
  */
-
 public class Coordinates {
-    private int coordX;
-    private int coordY;
+    private float coordX;
+    private float coordY;
 
-    public Coordinates(int x, int y) {
+    public Coordinates(float x, float y) {
         coordX = x;
         coordY = y;
     }
 
-    public int getCoordX() {
+    public float getCoordX() {
         return coordX;
     }
 
-    public int getCoordY() {
+    public float getCoordY() {
         return coordY;
     }
